issue-62: moviendo endpoint de trabajos a su proprio archivo, agregando suporte para conectar al nuevo backend para trabajos
